Владение объектом

+++ server/src/main/java/se/ifmo/blazingzephyr/commands/RemoveById.java
@@ -1,6 +1,8 @@
 package se.ifmo.blazingzephyr.commands;
 
 import java.sql.SQLException;
+import java.util.Optional;
+
 import se.ifmo.blazingzephyr.ServerContext;
 import se.ifmo.blazingzephyr.model.Organization;
 import se.ifmo.blazingzephyr.networking.CommandType;
@@ -26,24 +28,36 @@ public class RemoveById implements Command<WithId> {
      * {@inheritDoc}
      */
     @Override
-    public String execute(ServerContext ctx, WithId args) throws IllegalArgumentException {
+    public String execute(ServerContext ctx, WithId args, String login) throws IllegalArgumentException {
 
         long id = args.id();
+        Optional<Organization> org;
         try {
+            org = ctx.database().selectById(id);
+        } catch (SQLException ex) {
+            return "Произошла ошибка во время получения объекта из базы данных: " + ex.getMessage();
+        }
+
+        if (org.isEmpty())
+        {
+            return "Организации с искомым ID не существует.";
+        }
+
+        if (!org.get().getOwner().equals(login))
+        {
+            return "Невозможно удалить объект, не принадлежащий данному пользователю.";
+        }
+
+        try {
+
+
             boolean success = ctx.database().deleteById(id);
             if (!success)
             {
-                return "Организация не была удалена. " +
-                    "Проверьте существование организации с искомым ID.";
+                return "Организация не была удалена.";
             }
             else
             {
-                java.util.Optional<Organization> org = ctx
-                    .collection()
-                    .stream()
-                    .filter(o -> o.getId() == args.id())
-                    .findFirst();
-                
                 ctx.collection().remove(org.get());
                 return String.format("Организация с ID %d успешно удалена.%n", id);
             }
